Pin the cascade's scan dtypes so complex64 survives x64
The cluster failure is a defect in s5/modal_prospective.py, not in the test.
With JAX_ENABLE_X64=1 an unqualified np.ones(...) is float64, so

    elements = lambda_bar * np.ones((length, modes))

promoted a complex64 lambda_bar to complex128 while the complex64 drive
stayed as it was, and associative_scan raised

    lax.concatenate requires arguments to have the same dtypes,
    got complex64, complex128

That is exactly the production mix: the model runs in float32 while the
tests enable x64. The first seventeen tests passed because they build their
fixtures at the session dtype, so both operands promoted together; the
production-shape test pins complex64 deliberately and caught it.

Fix, in this branch's own module only:

  * native_states builds its broadcast array with dtype=lambda_bar.dtype and
    casts the drive to the same dtype, so both scan operands are pinned;
  * apply_stage casts its assembled drive to the sequence dtype for the same
    reason (its poles array was already qualified);
  * added_poles returns a float32 empty array rather than an untyped one;
  * the layer's parameters are created with dtype=np.float32 explicitly, as
    the other arms' are -- an unqualified np.full is float64 under x64 and
    would promote the whole cascade.

s5/ssm.py is NOT touched: it carries the same unqualified np.ones pattern,
but it is Native S5 and must stay byte-identical, and production never runs
it with x64 enabled. Verified identical, along with seq_model.py and
train_helpers.py.

Regression added: test_complex64_inputs_survive_x64_being_enabled runs
native_states, one stage, the cascade and both orientations with a complex64
fixture while x64 is on, asserting complex64 out of each, and separately
asserts that a complex128 fixture stays complex128 -- the fix pins dtypes,
it does not force single precision.

+++ s5/modal_prospective_ssm.py
@@ -62,21 +62,25 @@ class ModalProspectiveS5SSM(S5SSM):
     def setup(self):
         super().setup()
         shape = (self.P,)
+        # parameters are float32 explicitly, as the other arms' are: an
+        # unqualified np.full is float64 whenever x64 is enabled, which
+        # would promote the whole cascade and break the scan's dtypes
         raw_d = float(np.log(np.expm1(self.d_init)))
         raw_delta = float(np.log(np.expm1(self.delta_init)))
         self.stage_d_raw = [
             self.param(f"stage{index}_d_raw",
-                       lambda rng, s, value=raw_d: np.full(s, value), shape)
+                       lambda rng, s, value=raw_d:
+                       np.full(s, value, dtype=np.float32), shape)
             for index in range(self.stages)]
         self.stage_delta_raw = [
             self.param(f"stage{index}_delta_raw",
-                       lambda rng, s, value=raw_delta: np.full(s, value),
-                       shape)
+                       lambda rng, s, value=raw_delta:
+                       np.full(s, value, dtype=np.float32), shape)
             for index in range(self.stages)]
         self.stage_gate_raw = [
             self.param(f"stage{index}_gate_raw",
                        lambda rng, s, value=self.gate_init:
-                       np.full(s, value), shape)
+                       np.full(s, value, dtype=np.float32), shape)
             for index in range(self.stages)]
         b_tilde = self.B[..., 0] + 1j * self.B[..., 1]
         step = self.step_rescale * np.exp(self.log_step[:, 0])
